genericize the test executor
