+++ src/test/java/org/checkerframework/specimin/NoDependenciesReturnsSameTest.java
@@ -1,15 +1,6 @@
 package org.checkerframework.specimin;
 
-import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.concurrent.Executors;
-import java.util.function.Consumer;
-import org.junit.Assert;
 import org.junit.Test;
 
 /**
@@ -18,83 +9,7 @@ import org.junit.Test;
  */
 public class NoDependenciesReturnsSameTest {
   @Test
-  @SuppressWarnings("all")
   public void runTest() throws IOException {
-    // Create output directory
-    Path outputDir = null;
-    try {
-      outputDir = Files.createTempDirectory("specimin-test-");
-    } catch (IOException e) {
-      // TODO: what should we do if no temp directory can be created?
-      Assert.fail("cannot create temporary directory for test output: " + e);
-      return;
-    }
-    if (outputDir == null) {
-      Assert.fail("temporary directory for output was null");
-      return;
-    }
-    // Run specimin on target
-    SpeciminRunner.main(
-        "--outputDirectory",
-        outputDir.toAbsolutePath().toString(),
-        "--root",
-        Path.of("src/test/resources/nodependenciesreturnssame/input/").toAbsolutePath().toString(),
-        "--targetFiles",
-        "com/example/Simple.java");
-
-    // Diff the files to ensure that specimin's output is what we expect
-    ProcessBuilder builder = new ProcessBuilder();
-    boolean isWindows = System.getProperty("os.name").toLowerCase().startsWith("windows");
-    if (isWindows) {
-      // TODO: make this work
-      Assert.fail("specimin cannot be tested on Windows");
-      return;
-    } else {
-      builder.command(
-          "diff",
-          "-r",
-          outputDir.toAbsolutePath().toString(),
-          Path.of("src/test/resources/nodependenciesreturnssame/expected")
-              .toAbsolutePath()
-              .toString());
-    }
-    builder.directory(new File(System.getProperty("user.home")));
-    Process process = null;
-    try {
-      process = builder.start();
-    } catch (IOException e) {
-      Assert.fail("cannot start diff process: " + e);
-      return;
-    }
-    StringBuilder processOutput = new StringBuilder();
-    StreamGobbler streamGobbler =
-        new StreamGobbler(process.getInputStream(), s -> processOutput.append(s));
-    Executors.newSingleThreadExecutor().submit(streamGobbler);
-    int exitCode = 0;
-    try {
-      exitCode = process.waitFor();
-    } catch (InterruptedException e) {
-      Assert.fail("diff process interrupted: " + e);
-      return;
-    }
-    Assert.assertEquals(
-        "Diff failed with the following output: " + processOutput.toString() + "\n Error codes: ",
-        0,
-        exitCode);
-  }
-
-  private static class StreamGobbler implements Runnable {
-    private InputStream inputStream;
-    private Consumer<String> consumer;
-
-    public StreamGobbler(InputStream inputStream, Consumer<String> consumer) {
-      this.inputStream = inputStream;
-      this.consumer = consumer;
-    }
-
-    @Override
-    public void run() {
-      new BufferedReader(new InputStreamReader(inputStream)).lines().forEach(consumer);
-    }
+    SpeciminTestExecutor.runTest("nodependenciesreturnssame");
   }
 }
